use transformParamsToQuery
