+++ packages/core/utils/lib/convert-query-params.js
@@ -6,10 +6,22 @@
  * Converts the standard Strapi REST query params to a more usable format for querying
  * You can read more here: https://docs.strapi.io/developer-docs/latest/developer-resources/database-apis-reference/rest-api.html#filters
  */
-const { has, isEmpty, isObject, isPlainObject, cloneDeep, get, mergeAll } = require('lodash/fp');
+const {
+  has,
+  isEmpty,
+  isObject,
+  isPlainObject,
+  cloneDeep,
+  get,
+  mergeAll,
+  isNil,
+  toNumber,
+  isInteger,
+} = require('lodash/fp');
 const _ = require('lodash');
 const parseType = require('./parse-type');
 const contentTypesUtils = require('./content-types');
+const { PaginationError } = require('./errors');
 
 const { PUBLISHED_AT_ATTRIBUTE } = contentTypesUtils.constants;
 
@@ -389,6 +401,80 @@ const convertPublicationStateParams = (type, params = {}, query = {}) => {
   }
 };
 
+const transformParamsToQuery = (uid, params) => {
+  // NOTE: can be a CT, a Compo or nothing in the case of polymorphism (DZ & morph relations)
+  const schema = strapi.getModel(uid);
+
+  const query = {};
+
+  const { _q, sort, filters, fields, populate, page, pageSize, start, limit } = params;
+
+  if (!isNil(_q)) {
+    query._q = _q;
+  }
+
+  if (!isNil(sort)) {
+    query.orderBy = convertSortQueryParams(sort);
+  }
+
+  if (!isNil(filters)) {
+    query.where = convertFiltersQueryParams(filters, schema);
+  }
+
+  if (!isNil(fields)) {
+    query.select = convertFieldsQueryParams(fields);
+  }
+
+  if (!isNil(populate)) {
+    query.populate = convertPopulateQueryParams(populate, schema);
+  }
+
+  const isPagePagination = !isNil(page) || !isNil(pageSize);
+  const isOffsetPagination = !isNil(start) || !isNil(limit);
+
+  if (isPagePagination && isOffsetPagination) {
+    throw new PaginationError(
+      'Invalid pagination attributes. You cannot use page and offset pagination in the same query'
+    );
+  }
+
+  if (!isNil(page)) {
+    const pageVal = toNumber(page);
+
+    if (!isInteger(pageVal) || pageVal <= 0) {
+      throw new PaginationError(
+        `Invalid 'page' parameter. Expected an integer > 0, received: ${page}`
+      );
+    }
+
+    query.page = pageVal;
+  }
+
+  if (!isNil(pageSize)) {
+    const pageSizeVal = toNumber(pageSize);
+
+    if (!isInteger(pageSizeVal) || pageSizeVal <= 0) {
+      throw new PaginationError(
+        `Invalid 'pageSize' parameter. Expected an integer > 0, received: ${page}`
+      );
+    }
+
+    query.pageSize = pageSizeVal;
+  }
+
+  if (!isNil(start)) {
+    query.offset = convertStartQueryParams(start);
+  }
+
+  if (!isNil(limit)) {
+    query.limit = convertLimitQueryParams(limit);
+  }
+
+  convertPublicationStateParams(schema, params, query);
+
+  return query;
+};
+
 module.exports = {
   convertSortQueryParams,
   convertStartQueryParams,
@@ -397,4 +483,5 @@ module.exports = {
   convertFiltersQueryParams,
   convertFieldsQueryParams,
   convertPublicationStateParams,
+  transformParamsToQuery,
 };
